Refactor validation.js: move to js/utils/validation and split into modules to enhance code maintainability and scalability
- Each validation concern is now handled by its own module
