+++ pages/reservation.html
@@ -56,21 +56,18 @@
                   <div class="error-message"></div>
                   <input
                     type="text"
-                    id="date-month"
                     class="form-input date-input"
                     name="date_month"
                     placeholder="MM"
                   />
                   <input
                     type="text"
-                    id="date-day"
                     class="form-input date-input"
                     name="date_day"
                     placeholder="DD"
                   />
                   <input
                     type="text"
-                    id="date-year"
                     class="form-input input-large date-input"
                     name="date_year"
                     placeholder="YYYY"
@@ -83,14 +80,12 @@
                   <div class="error-message"></div>
                   <input
                     type="text"
-                    id="time_hour"
                     class="form-input time-input"
                     name="time_hour"
                     placeholder="09"
                   />
                   <input
                     type="text"
-                    id="time_minute"
                     class="form-input time-input"
                     name="time_minute"
                     placeholder="00"
@@ -98,7 +93,7 @@
                   <select
                     class="form-input input-large"
                     name="am_pm"
-                    id="am_pm"
+                    id="am-pm"
                   >
                     <option value="am">AM</option>
                     <option value="pm">PM</option>
